+++ contracts/University.sol
@@ -9,29 +9,45 @@ import "./Classroom.sol";
 import "./Student.sol";
 import "./StudentApplication.sol";
 
+
 interface CERC20 {
     function mint(uint256) external returns (uint256);
+
     function exchangeRateCurrent() external returns (uint256);
+
     function supplyRatePerBlock() external returns (uint256);
-    function redeem(uint) external returns (uint);
-    function redeemUnderlying(uint) external returns (uint);
-    function getCash() external returns (uint);
-    function balanceOfUnderlying(address account) external returns (uint);
+
+    function redeem(uint256) external returns (uint256);
+
+    function redeemUnderlying(uint256) external returns (uint256);
+
+    function getCash() external returns (uint256);
+
+    function balanceOfUnderlying(address account) external returns (uint256);
 }
+
 
 contract University is Ownable, AccessControl {
     using SafeMath for uint256;
 
     //CLASSLIST_ADMIN_ROLE can add new manually created classes to the list
-    bytes32 public constant CLASSLIST_ADMIN_ROLE = keccak256("CLASSLIST_ADMIN_ROLE");
+    bytes32 public constant CLASSLIST_ADMIN_ROLE = keccak256(
+        "CLASSLIST_ADMIN_ROLE"
+    );
     // FUNDS_MANAGER_ROLE can withdraw funds from the contract
-    bytes32 public constant FUNDS_MANAGER_ROLE = keccak256("FUNDS_MANAGER_ROLE");
+    bytes32 public constant FUNDS_MANAGER_ROLE = keccak256(
+        "FUNDS_MANAGER_ROLE"
+    );
     // GRANTS_MANAGER_ROLE can approve/decline grant claims
-    bytes32 public constant GRANTS_MANAGER_ROLE = keccak256("GRANTS_MANAGER_ROLE");
+    bytes32 public constant GRANTS_MANAGER_ROLE = keccak256(
+        "GRANTS_MANAGER_ROLE"
+    );
     // CLASSROOM_ROLE can manage itself inside the University and registering student applications
     bytes32 public constant CLASSROOM_ROLE = keccak256("CLASSROOM_ROLE");
     // READ_STUDENT_LIST_ROLE allow reading students list
-    bytes32 public constant READ_STUDENT_LIST_ROLE = keccak256("READ_STUDENT_LIST_ROLE");
+    bytes32 public constant READ_STUDENT_LIST_ROLE = keccak256(
+        "READ_STUDENT_LIST_ROLE"
+    );
     // STUDENT_ROLE allow asking for grants and requesting a classroom from a successful application
     bytes32 public constant STUDENT_ROLE = keccak256("STUDENT_ROLE");
 
@@ -51,7 +67,12 @@ contract University is Ownable, AccessControl {
     CERC20 public cToken;
     IERC20 public daiToken;
 
-    constructor(bytes32 _name, uint24 _cut, address daiAddress, address compoundAddress) public {
+    constructor(
+        bytes32 _name,
+        uint24 _cut,
+        address daiAddress,
+        address compoundAddress
+    ) public {
         name = _name;
         cut = _cut;
         _classList = new Classroom[](0);
@@ -81,23 +102,64 @@ contract University is Ownable, AccessControl {
         return hasRole(CLASSROOM_ROLE, classroom);
     }
 
-    function studentIsRegistered(address student) public view returns (bool){
-        require(hasRole(READ_STUDENT_LIST_ROLE, _msgSender()), "University: caller doesn't have READ_STUDENT_LIST_ROLE");
+    function studentIsRegistered(address student) public view returns (bool) {
+        require(
+            hasRole(READ_STUDENT_LIST_ROLE, _msgSender()),
+            "University: caller doesn't have READ_STUDENT_LIST_ROLE"
+        );
         return hasRole(STUDENT_ROLE, student);
     }
 
     //ex: owner, name, 0.2 * 10**6, 0.5 * 10**6, 0, 50 * (10 ** 18), 30 days, challengeAddress
-    function newClassRoom(address owner, bytes32 cName, uint24 cCut, uint24 cPCut,
-            int32 minScore, uint entryPrice, uint duration, address challengeAddress) public {
-        require(hasRole(CLASSLIST_ADMIN_ROLE, _msgSender()), "University: caller doesn't have CLASSLIST_ADMIN_ROLE");
-        _newClassRoom(owner, cName, cCut, cPCut, minScore, entryPrice, duration, challengeAddress);
+    function newClassRoom(
+        address owner,
+        bytes32 cName,
+        uint24 cCut,
+        uint24 cPCut,
+        int32 minScore,
+        uint256 entryPrice,
+        uint256 duration,
+        address challengeAddress
+    ) public {
+        require(
+            hasRole(CLASSLIST_ADMIN_ROLE, _msgSender()),
+            "University: caller doesn't have CLASSLIST_ADMIN_ROLE"
+        );
+        _newClassRoom(
+            owner,
+            cName,
+            cCut,
+            cPCut,
+            minScore,
+            entryPrice,
+            duration,
+            challengeAddress
+        );
     }
 
-    function _newClassRoom(address owner, bytes32 cName, uint24 cCut, uint24 cPCut,
-            int32 minScore, uint entryPrice, uint duration, address challengeAddress) internal {
+    function _newClassRoom(
+        address owner,
+        bytes32 cName,
+        uint24 cCut,
+        uint24 cPCut,
+        int32 minScore,
+        uint256 entryPrice,
+        uint256 duration,
+        address challengeAddress
+    ) internal {
         //TODO: fetch contract from external factory to reduce size
-        Classroom classroom = new Classroom(cName, cCut, cPCut, minScore, entryPrice, duration,
-            address(this), address(daiToken), address(cToken), challengeAddress);
+        Classroom classroom = new Classroom(
+            cName,
+            cCut,
+            cPCut,
+            minScore,
+            entryPrice,
+            duration,
+            address(this),
+            address(daiToken),
+            address(cToken),
+            challengeAddress
+        );
         classroom.transferOwnership(owner);
         _classList.push(classroom);
         grantRole(READ_STUDENT_LIST_ROLE, address(classroom));
@@ -107,7 +169,10 @@ contract University is Ownable, AccessControl {
 
     //TODO: Use GSN to improve UX for new student
     function studentSelfRegister(bytes32 sName) public {
-        require(_studentApplicationsMapping[_msgSender()].length == 0, "University: student already registered");
+        require(
+            _studentApplicationsMapping[_msgSender()].length == 0,
+            "University: student already registered"
+        );
         _newStudent(sName, _msgSender());
     }
 
@@ -121,8 +186,13 @@ contract University is Ownable, AccessControl {
         grantRole(STUDENT_ROLE, address(student));
     }
 
-    function registerStudentApplication(address student, address application) public {
-        require(hasRole(CLASSROOM_ROLE, _msgSender()), "University: caller doesn't have CLASSROOM_ROLE");
+    function registerStudentApplication(address student, address application)
+        public
+    {
+        require(
+            hasRole(CLASSROOM_ROLE, _msgSender()),
+            "University: caller doesn't have CLASSROOM_ROLE"
+        );
         _studentApplicationsMapping[student].push(application);
     }
 
@@ -130,57 +200,122 @@ contract University is Ownable, AccessControl {
         return viewStudentApplications(_msgSender());
     }
 
-    function viewStudentApplications(address addr) public view returns (address[] memory) {
-        require(addr == _msgSender() || hasRole(GRANTS_MANAGER_ROLE, _msgSender()), "Classroom: read permission denied");
+    function viewStudentApplications(address addr)
+        public
+        view
+        returns (address[] memory)
+    {
+        require(
+            addr == _msgSender() || hasRole(GRANTS_MANAGER_ROLE, _msgSender()),
+            "Classroom: read permission denied"
+        );
         return _studentApplicationsMapping[addr];
     }
 
-    function studentRequestClassroom(address applicationAddr,
-            bytes32 cName, uint24 cCut, uint24 cPCut, int32 minScore, uint entryPrice, uint duration, address challenge) public {
-        require(hasRole(STUDENT_ROLE, _msgSender()), "University: caller doesn't have STUDENT_ROLE");
+    function studentRequestClassroom(
+        address applicationAddr,
+        bytes32 cName,
+        uint24 cCut,
+        uint24 cPCut,
+        int32 minScore,
+        uint256 entryPrice,
+        uint256 duration,
+        address challenge
+    ) public {
+        require(
+            hasRole(STUDENT_ROLE, _msgSender()),
+            "University: caller doesn't have STUDENT_ROLE"
+        );
         StudentApplication application = StudentApplication(applicationAddr);
-        require(checkForStudentApplication(_msgSender(), applicationAddr), "University: caller is not student of this application");
-        require(application.applicationState() == 3, "University: application is not successful");
-        _newClassRoom(Student(_msgSender()).owner(), cName, cCut, cPCut, minScore, entryPrice, duration, challenge);
+        require(
+            checkForStudentApplication(_msgSender(), applicationAddr),
+            "University: caller is not student of this application"
+        );
+        require(
+            application.applicationState() == 3,
+            "University: application is not successful"
+        );
+        _newClassRoom(
+            Student(_msgSender()).owner(),
+            cName,
+            cCut,
+            cPCut,
+            minScore,
+            entryPrice,
+            duration,
+            challenge
+        );
     }
 
-    function checkForStudentApplication(address studentAddress, address applicationAddress) internal view returns (bool) {
-        for (uint i = 0; i < _studentApplicationsMapping[studentAddress].length ; i++) {
-            if (_studentApplicationsMapping[studentAddress][i] == applicationAddress) return true;
+    function checkForStudentApplication(
+        address studentAddress,
+        address applicationAddress
+    ) internal view returns (bool) {
+        for (
+            uint256 i = 0;
+            i < _studentApplicationsMapping[studentAddress].length;
+            i++
+        ) {
+            if (
+                _studentApplicationsMapping[studentAddress][i] ==
+                applicationAddress
+            ) return true;
         }
         return false;
     }
 
-    function addStudentScore(address student, int32 val) public  {
-        require(hasRole(CLASSROOM_ROLE, _msgSender()), "University: caller doesn't have CLASSROOM_ROLE");
+    function addStudentScore(address student, int32 val) public {
+        require(
+            hasRole(CLASSROOM_ROLE, _msgSender()),
+            "University: caller doesn't have CLASSROOM_ROLE"
+        );
         Student(student).addScore(val);
     }
 
-    function subStudentScore(address student,int32 val) public  {
-        require(hasRole(CLASSROOM_ROLE, _msgSender()), "University: caller doesn't have CLASSROOM_ROLE");
+    function subStudentScore(address student, int32 val) public {
+        require(
+            hasRole(CLASSROOM_ROLE, _msgSender()),
+            "University: caller doesn't have CLASSROOM_ROLE"
+        );
         Student(student).subScore(val);
     }
 
-    function applyFunds(uint val) public {
-        require(hasRole(FUNDS_MANAGER_ROLE, _msgSender()), "University: caller doesn't have FUNDS_MANAGER_ROLE");
+    function applyFunds(uint256 val) public {
+        require(
+            hasRole(FUNDS_MANAGER_ROLE, _msgSender()),
+            "University: caller doesn't have FUNDS_MANAGER_ROLE"
+        );
         daiToken.approve(address(cToken), val);
         cToken.mint(val);
     }
 
-    function recoverFunds(uint val) public {
-        require(hasRole(FUNDS_MANAGER_ROLE, _msgSender()), "University: caller doesn't have FUNDS_MANAGER_ROLE");
+    function recoverFunds(uint256 val) public {
+        require(
+            hasRole(FUNDS_MANAGER_ROLE, _msgSender()),
+            "University: caller doesn't have FUNDS_MANAGER_ROLE"
+        );
         cToken.redeemUnderlying(val);
     }
 
-    function spendFunds(address to, uint val) public {
-        require(hasRole(FUNDS_MANAGER_ROLE, _msgSender()), "University: caller doesn't have FUNDS_MANAGER_ROLE");
+    function spendFunds(address to, uint256 val) public {
+        require(
+            hasRole(FUNDS_MANAGER_ROLE, _msgSender()),
+            "University: caller doesn't have FUNDS_MANAGER_ROLE"
+        );
         daiToken.transfer(to, val);
     }
 
-    function allowFunds(address to, uint val) public {
-        require(hasRole(FUNDS_MANAGER_ROLE, _msgSender()), "University: caller doesn't have FUNDS_MANAGER_ROLE");
+    function allowFunds(address to, uint256 val) public {
+        require(
+            hasRole(FUNDS_MANAGER_ROLE, _msgSender()),
+            "University: caller doesn't have FUNDS_MANAGER_ROLE"
+        );
         daiToken.approve(to, val);
     }
 
     //TODO: manage grants
+
+    //TODO: implement funds manager
+
+    //TODO: implement funds manager governance
 }
